base frontend setup, waiting on json photos and implementing person selector
