+++ frontend/assets/css/styles.css
@@ -1,0 +1,136 @@
+@font-face {
+    font-family: 'One Day';
+    src: url('https://storage.googleapis.com/hackgt/fonts/ONEDAY.ttf');
+}
+
+body {
+    overflow-y: hidden;
+    font-family: 'One Day', sans-serif;
+    word-spacing: 120%; /* Don't think this is working */
+}
+
+.home-dark {
+    height: 1000px;
+    background-size: cover;
+    position: relative
+}
+
+.section-wrap {
+    max-width: 1500px;
+    width: 90%;
+}
+
+.section-header {
+    padding: 18px;
+    border-bottom: 2px solid #434a52;
+}
+
+.section-title {
+    display: inline-block;
+    font-size: 1.25rem;
+}
+
+.section-body {
+    max-height: 1000px;
+    padding: 18px;
+}
+
+.section-footer {
+    padding: 18px;
+}
+
+.start-photo-container {
+    max-width: 320px;
+    width: 90%;
+    padding: 40px;
+    border-radius: 4px;
+    transform: translate(-50%, -50%);
+    position: absolute;
+    top: 30%;
+    left: 50%;
+    color: #fff;
+}
+
+/* .home-dark form .btn-primary { */
+.photo-btn {
+    background: #214a80;
+    border: none;
+    border-radius: 4px;
+    color: #fff;
+    cursor: pointer;
+    display: block;
+    padding: 11px;
+    box-shadow: none;
+    margin: 26px auto 0;
+    text-shadow: none;
+    outline: 0;
+    width: 300px;
+}
+
+/* TODO: put some color in here? */
+.photo-btn:active,
+.photo-btn:hover {
+    background: #214a80;
+    outline: 0;
+}
+
+.home-dark .illustration {
+    text-align: center;
+    padding: 15px 0 20px;
+    font-size: 100px;
+    color: #2980ef
+}
+
+.home-dark form .form-control {
+    background: #fff;
+    border: none;
+    border-bottom: 1px solid #434a52;
+    border-radius: 0;
+    box-shadow: none;
+    outline: 0;
+    color: inherit
+}
+
+.text-center {
+    color: black;
+    margin: 2em 0;
+    font-size: 5em;
+}
+
+.radio-container {
+    display: flex;
+    width: 80vw;
+    left: 50%;
+    transform: translateX(-50%);
+    position: absolute;
+    bottom: 150px;
+    /* overflow-x: scroll; */
+}
+
+input[type='radio'] {
+    visibility: hidden;
+}
+
+input[type='radio']:checked + label {
+    border: 10px black solid;
+}
+
+.countdown {
+    display: block;
+    margin: 10px auto;
+}
+
+.underline {
+    border-bottom: 2px solid black;
+}
+
+.filter-option:hover .underline, .filter-option.active .underline {
+    border-bottom: 2px solid white;
+}
+
+.restart-button {
+    width: 130px;
+    padding: 4px;
+    float: right;
+    margin: 0;
+}
