--- frontend/static/css/styles.css
@@ -1,123 +1,0 @@
-@font-face {
-    font-family: 'One Day';
-    src: url('https://storage.googleapis.com/hackgt/fonts/ONEDAY.ttf');
-}
-
-body {
-    overflow-y: hidden;
-    font-family: 'One Day', sans-serif;
-}
-
-.home-dark {
-    height: 1000px;
-    background-size: cover;
-    position: relative
-}
-
-.section-wrap {
-    max-width: 1500px;
-    width: 90%;
-}
-
-.section-header {
-    padding: 18px;
-    border-bottom: 2px solid #434a52;
-}
-
-.section-title {
-    font-size: 1.25rem;
-}
-
-.section-body {
-    max-height: 700px;
-    padding: 18px;
-}
-
-.section-footer {
-    padding: 18px;
-}
-
-.start-photo-container {
-    max-width: 320px;
-    width: 90%;
-    padding: 40px;
-    border-radius: 4px;
-    transform: translate(-50%, -50%);
-    position: absolute;
-    top: 30%;
-    left: 50%;
-    color: #fff;
-}
-
-/* .home-dark form .btn-primary { */
-.photo-btn {
-    background: #214a80;
-    border: none;
-    border-radius: 4px;
-    color: #fff;
-    cursor: pointer;
-    padding: 11px;
-    box-shadow: none;
-    margin-top: 26px;
-    text-shadow: none;
-    outline: 0;
-}
-
-/* TODO: put some color in here? */
-.photo-btn:active,
-.photo-btn:hover {
-    background: #214a80;
-    outline: 0;
-}
-
-
-.photo-take {
-    width: 300px;
-}
-
-
-
-.home-dark .illustration {
-    text-align: center;
-    padding: 15px 0 20px;
-    font-size: 100px;
-    color: #2980ef
-}
-
-.home-dark form .form-control {
-    background: #fff;
-    border: none;
-    border-bottom: 1px solid #434a52;
-    border-radius: 0;
-    box-shadow: none;
-    outline: 0;
-    color: inherit
-}
-
-.text-center {
-    color: black;
-    margin: 2em 0;
-    font-size: 5em;
-}
-
-.radio-container {
-    display: flex;
-    width: 80vw;
-    left: 50%;
-    transform: translateX(-50%);
-    position: absolute;
-    bottom: 150px;
-    /* overflow-x: scroll; */
-}
-
-input[type='radio'] {
-    visibility: hidden;
-}
-
-input[type='radio']:checked + label {
-    border: 10px black solid;
-}
-
-.countdown {
-    position: absolute;
-}
