--- frontend/static/js/izitoast.js
@@ -1,6 +1,0 @@
-/*
-* iziToast | v1.3.0
-* http://izitoast.marcelodolce.com
-* by Marcelo Dolce.
-*/
-!function(t,e){"function"==typeof define&&define.amd?define([],e(t)):"object"==typeof exports?module.exports=e(t):t.iziToast=e(t)}("undefined"!=typeof global?global:window||this.window||this.global,function(t){"use strict";var e={},n="iziToast",o=(document.querySelector("body"),!!/Mobi/.test(navigator.userAgent)),i=/Chrome/.test(navigator.userAgent)&&/Google Inc/.test(navigator.vendor),s="undefined"!=typeof InstallTrigger,a="ontouchstart"in document.documentElement,r=["bottomRight","bottomLeft","bottomCenter","topRight","topLeft","topCenter","center"],l={info:{color:"blue",icon:"ico-info"},success:{color:"green",icon:"ico-success"},warning:{color:"orange",icon:"ico-warning"},error:{color:"red",icon:"ico-error"},question:{color:"yellow",icon:"ico-question"}},d=568,c={};e.childrens={};var u={id:null,"class":"",title:"",titleColor:"",titleSize:"",titleLineHeight:"",message:"",messageColor:"",messageSize:"",messageLineHeight:"",backgroundColor:"",theme:"light",color:"",icon:"",iconText:"",iconColor:"",image:"",imageWidth:50,maxWidth:null,zindex:null,layout:1,balloon:!1,close:!0,closeOnEscape:!1,closeOnClick:!1,rtl:!1,position:"bottomRight",target:"",targetFirst:!0,toastOnce:!1,timeout:5e3,animateInside:!0,drag:!0,pauseOnHover:!0,resetOnHover:!1,progressBar:!0,progressBarColor:"",progressBarEasing:"linear",overlay:!1,overlayClose:!1,overlayColor:"rgba(0, 0, 0, 0.6)",transitionIn:"fadeInUp",transitionOut:"fadeOut",transitionInMobile:"fadeInUp",transitionOutMobile:"fadeOutDown",buttons:{},inputs:{},onOpening:function(){},onOpened:function(){},onClosing:function(){},onClosed:function(){}};if("remove"in Element.prototype||(Element.prototype.remove=function(){this.parentNode&&this.parentNode.removeChild(this)}),"function"!=typeof window.CustomEvent){var p=function(t,e){e=e||{bubbles:!1,cancelable:!1,detail:void 0};var n=document.createEvent("CustomEvent");return n.initCustomEvent(t,e.bubbles,e.cancelable,e.detail),n};p.prototype=window.Event.prototype,window.CustomEvent=p}var m=function(t,e,n){if("[object Object]"===Object.prototype.toString.call(t))for(var o in t)Object.prototype.hasOwnProperty.call(t,o)&&e.call(n,t[o],o,t);else if(t)for(var i=0,s=t.length;s>i;i++)e.call(n,t[i],i,t)},g=function(t,e){var n={};return m(t,function(e,o){n[o]=t[o]}),m(e,function(t,o){n[o]=e[o]}),n},f=function(t){var e=document.createDocumentFragment(),n=document.createElement("div");for(n.innerHTML=t;n.firstChild;)e.appendChild(n.firstChild);return e},v=function(t){return"#"==t.substring(0,1)||"rgb"==t.substring(0,3)||"hsl"==t.substring(0,3)},h=function(t){try{return btoa(atob(t))==t}catch(e){return!1}},y=function(){return{move:function(t,e,o,a){var r,l=.3,d=180;0!==a&&(t.classList.add(n+"-dragged"),t.style.transform="translateX("+a+"px)",a>0?(r=(d-a)/d,l>r&&e.hide(g(o,{transitionOut:"fadeOutRight",transitionOutMobile:"fadeOutRight"}),t,"drag")):(r=(d+a)/d,l>r&&e.hide(g(o,{transitionOut:"fadeOutLeft",transitionOutMobile:"fadeOutLeft"}),t,"drag")),t.style.opacity=r,l>r&&((i||s)&&(t.style.left=a+"px"),t.parentNode.style.opacity=l,this.stopMoving(t,null)))},startMoving:function(t,e,n,o){o=o||window.event;var i=a?o.touches[0].clientX:o.clientX,s=t.style.transform.replace("px)","");s=s.replace("translateX(","");var r=i-s;t.classList.remove(n.transitionIn),t.classList.remove(n.transitionInMobile),t.style.transition="",a?document.ontouchmove=function(o){o.preventDefault(),o=o||window.event;var i=o.touches[0].clientX,s=i-r;y.move(t,e,n,s)}:document.onmousemove=function(o){o.preventDefault(),o=o||window.event;var i=o.clientX,s=i-r;y.move(t,e,n,s)}},stopMoving:function(t,e){a?document.ontouchmove=function(){}:document.onmousemove=function(){},t.style.opacity="",t.style.transform="",t.classList.contains(n+"-dragged")&&(t.classList.remove(n+"-dragged"),t.style.transition="transform 0.4s ease, opacity 0.4s ease",setTimeout(function(){t.style.transition=""},400))}}}();return e.setSetting=function(t,n,o){e.childrens[t][n]=o},e.getSetting=function(t,n){return e.childrens[t][n]},e.destroy=function(){m(document.querySelectorAll("."+n+"-wrapper"),function(t,e){t.remove()}),m(document.querySelectorAll("."+n),function(t,e){t.remove()}),document.removeEventListener(n+"-opened",{},!1),document.removeEventListener(n+"-opening",{},!1),document.removeEventListener(n+"-closing",{},!1),document.removeEventListener(n+"-closed",{},!1),document.removeEventListener("keyup",{},!1),c={}},e.settings=function(t){e.destroy(),c=t,u=g(u,t||{})},m(l,function(t,n){e[n]=function(e){var n=g(c,e||{});n=g(t,n||{}),this.show(n)}}),e.progress=function(t,e,o){var i=this,s=e.getAttribute("data-iziToast-ref"),a=g(this.childrens[s],t||{}),r=e.querySelector("."+n+"-progressbar div");return{start:function(){"undefined"==typeof a.time.REMAINING&&(e.classList.remove(n+"-reseted"),null!==r&&(r.style.transition="width "+a.timeout+"ms "+a.progressBarEasing,r.style.width="0%"),a.time.START=(new Date).getTime(),a.time.END=a.time.START+a.timeout,a.time.TIMER=setTimeout(function(){clearTimeout(a.time.TIMER),e.classList.contains(n+"-closing")||(i.hide(a,e,"timeout"),"function"==typeof o&&o.apply(i))},a.timeout),i.setSetting(s,"time",a.time))},pause:function(){if("undefined"!=typeof a.time.START&&!e.classList.contains(n+"-paused")&&!e.classList.contains(n+"-reseted")){if(e.classList.add(n+"-paused"),a.time.REMAINING=a.time.END-(new Date).getTime(),clearTimeout(a.time.TIMER),i.setSetting(s,"time",a.time),null!==r){var t=window.getComputedStyle(r),l=t.getPropertyValue("width");r.style.transition="none",r.style.width=l}"function"==typeof o&&setTimeout(function(){o.apply(i)},10)}},resume:function(){"undefined"!=typeof a.time.REMAINING?(e.classList.remove(n+"-paused"),null!==r&&(r.style.transition="width "+a.time.REMAINING+"ms "+a.progressBarEasing,r.style.width="0%"),a.time.END=(new Date).getTime()+a.time.REMAINING,a.time.TIMER=setTimeout(function(){clearTimeout(a.time.TIMER),e.classList.contains(n+"-closing")||(i.hide(a,e,"timeout"),"function"==typeof o&&o.apply(i))},a.time.REMAINING),i.setSetting(s,"time",a.time)):this.start()},reset:function(){clearTimeout(a.time.TIMER),delete a.time.REMAINING,i.setSetting(s,"time",a.time),e.classList.add(n+"-reseted"),e.classList.remove(n+"-paused"),null!==r&&(r.style.transition="none",r.style.width="100%"),"function"==typeof o&&setTimeout(function(){o.apply(i)},10)}}},e.hide=function(t,e,i){var s=this,a=g(this.childrens[e.getAttribute("data-iziToast-ref")],t||{});a.closedBy=i||null,delete a.time.REMAINING,"object"!=typeof e&&(e=document.querySelector(e)),e.classList.add(n+"-closing"),function(){var t=document.querySelector("."+n+"-overlay");if(null!==t){var e=t.getAttribute("data-iziToast-ref");e=e.split(",");var o=e.indexOf(String(a.ref));-1!==o&&e.splice(o,1),t.setAttribute("data-iziToast-ref",e.join()),0===e.length&&(t.classList.remove("fadeIn"),t.classList.add("fadeOut"),setTimeout(function(){t.remove()},700))}}(),(a.transitionIn||a.transitionInMobile)&&(e.classList.remove(a.transitionIn),e.classList.remove(a.transitionInMobile)),o||window.innerWidth<=d?a.transitionOutMobile&&e.classList.add(a.transitionOutMobile):a.transitionOut&&e.classList.add(a.transitionOut);var r=e.parentNode.offsetHeight;e.parentNode.style.height=r+"px",e.style.pointerEvents="none",(!o||window.innerWidth>d)&&(e.parentNode.style.transitionDelay="0.2s");try{var l=new CustomEvent(n+"-closing",{detail:a,bubbles:!0,cancelable:!0});document.dispatchEvent(l)}catch(c){console.warn(c)}setTimeout(function(){e.parentNode.style.height="0px",e.parentNode.style.overflow="",setTimeout(function(){delete s.childrens[a.ref],e.parentNode.remove();try{var t=new CustomEvent(n+"-closed",{detail:a,bubbles:!0,cancelable:!0});document.dispatchEvent(t)}catch(o){console.warn(o)}"undefined"!=typeof a.onClosed&&a.onClosed.apply(null,[a,e,i])},1e3)},200),"undefined"!=typeof a.onClosing&&a.onClosing.apply(null,[a,e,i])},e.show=function(t){var i=this,s=g(c,t||{});if(s=g(u,s),s.time={},s.toastOnce&&s.id&&document.querySelectorAll("."+n+"#"+s.id).length>0)return!1;s.ref=(new Date).getTime()+Math.floor(1e7*Math.random()+1),e.childrens[s.ref]=s;var l={body:document.querySelector("body"),overlay:document.createElement("div"),toast:document.createElement("div"),toastBody:document.createElement("div"),toastTexts:document.createElement("div"),toastCapsule:document.createElement("div"),icon:document.createElement("i"),cover:document.createElement("div"),buttons:document.createElement("div"),inputs:document.createElement("div"),wrapper:null};l.toast.setAttribute("data-iziToast-ref",s.ref),l.toast.appendChild(l.toastBody),l.toastCapsule.appendChild(l.toast),function(){if(l.toast.classList.add(n),l.toast.classList.add(n+"-opening"),l.toastCapsule.classList.add(n+"-capsule"),l.toastBody.classList.add(n+"-body"),l.toastTexts.classList.add(n+"-texts"),o||window.innerWidth<=d?s.transitionInMobile&&l.toast.classList.add(s.transitionInMobile):s.transitionIn&&l.toast.classList.add(s.transitionIn),s["class"]){var t=s["class"].split(" ");m(t,function(t,e){l.toast.classList.add(t)})}s.id&&(l.toast.id=s.id),s.rtl&&(l.toast.classList.add(n+"-rtl"),l.toast.setAttribute("dir","rtl")),s.layout>1&&l.toast.classList.add(n+"-layout"+s.layout),s.balloon&&l.toast.classList.add(n+"-balloon"),s.maxWidth&&(isNaN(s.maxWidth)?l.toast.style.maxWidth=s.maxWidth:l.toast.style.maxWidth=s.maxWidth+"px"),""===s.theme&&"light"===s.theme||l.toast.classList.add(n+"-theme-"+s.theme),s.color&&(v(s.color)?l.toast.style.background=s.color:l.toast.classList.add(n+"-color-"+s.color)),s.backgroundColor&&(l.toast.style.background=s.backgroundColor,s.balloon&&(l.toast.style.borderColor=s.backgroundColor))}(),function(){s.image&&(l.cover.classList.add(n+"-cover"),l.cover.style.width=s.imageWidth+"px",h(s.image.replace(/ /g,""))?l.cover.style.backgroundImage="url(data:image/png;base64,"+s.image.replace(/ /g,"")+")":l.cover.style.backgroundImage="url("+s.image+")",s.rtl?l.toastBody.style.marginRight=s.imageWidth+10+"px":l.toastBody.style.marginLeft=s.imageWidth+10+"px",l.toast.appendChild(l.cover))}(),function(){s.close?(l.buttonClose=document.createElement("button"),l.buttonClose.classList.add(n+"-close"),l.buttonClose.addEventListener("click",function(t){t.target;i.hide(s,l.toast,"button")}),l.toast.appendChild(l.buttonClose)):s.rtl?l.toast.style.paddingLeft="18px":l.toast.style.paddingRight="18px"}(),function(){s.progressBar&&(l.progressBar=document.createElement("div"),l.progressBarDiv=document.createElement("div"),l.progressBar.classList.add(n+"-progressbar"),l.progressBarDiv.style.background=s.progressBarColor,l.progressBar.appendChild(l.progressBarDiv),l.toast.appendChild(l.progressBar)),s.timeout&&(s.pauseOnHover&&!s.resetOnHover&&(l.toast.addEventListener("mouseenter",function(t){i.progress(s,l.toast).pause()}),l.toast.addEventListener("mouseleave",function(t){i.progress(s,l.toast).resume()})),s.resetOnHover&&(l.toast.addEventListener("mouseenter",function(t){i.progress(s,l.toast).reset()}),l.toast.addEventListener("mouseleave",function(t){i.progress(s,l.toast).start()})))}(),function(){s.icon&&(l.icon.setAttribute("class",n+"-icon "+s.icon),s.iconText&&l.icon.appendChild(document.createTextNode(s.iconText)),s.rtl?l.toastBody.style.paddingRight="33px":l.toastBody.style.paddingLeft="33px",s.iconColor&&(l.icon.style.color=s.iconColor),l.toastBody.appendChild(l.icon))}(),function(){s.title.length>0&&(l.strong=document.createElement("strong"),l.strong.classList.add(n+"-title"),l.strong.appendChild(f(s.title)),l.toastTexts.appendChild(l.strong),s.titleColor&&(l.strong.style.color=s.titleColor),s.titleSize&&(isNaN(s.titleSize)?l.strong.style.fontSize=s.titleSize:l.strong.style.fontSize=s.titleSize+"px"),s.titleLineHeight&&(isNaN(s.titleSize)?l.strong.style.lineHeight=s.titleLineHeight:l.strong.style.lineHeight=s.titleLineHeight+"px")),s.message.length>0&&(l.p=document.createElement("p"),l.p.classList.add(n+"-message"),l.p.appendChild(f(s.message)),l.toastTexts.appendChild(l.p),s.messageColor&&(l.p.style.color=s.messageColor),s.messageSize&&(isNaN(s.titleSize)?l.p.style.fontSize=s.messageSize:l.p.style.fontSize=s.messageSize+"px"),s.messageLineHeight&&(isNaN(s.titleSize)?l.p.style.lineHeight=s.messageLineHeight:l.p.style.lineHeight=s.messageLineHeight+"px")),s.title.length>0&&s.message.length>0&&(s.rtl?l.strong.style.marginLeft="10px":2===s.layout||s.rtl||(l.strong.style.marginRight="10px"))}(),l.toastBody.appendChild(l.toastTexts);var p;!function(){s.inputs.length>0&&(l.inputs.classList.add(n+"-inputs"),m(s.inputs,function(t,e){l.inputs.appendChild(f(t[0])),p=l.inputs.childNodes,p[e].classList.add(n+"-inputs-child"),t[3]&&setTimeout(function(){p[e].focus()},300),p[e].addEventListener(t[1],function(e){var n=t[2];return n(i,l.toast,this,e)})}),l.toastBody.appendChild(l.inputs))}(),function(){s.buttons.length>0&&(l.buttons.classList.add(n+"-buttons"),m(s.buttons,function(t,e){l.buttons.appendChild(f(t[0]));var o=l.buttons.childNodes;o[e].classList.add(n+"-buttons-child"),t[2]&&setTimeout(function(){o[e].focus()},300),o[e].addEventListener("click",function(e){e.preventDefault();var n=t[1];return n(i,l.toast,this,e,p)})})),l.toastBody.appendChild(l.buttons)}(),s.message.length>0&&(s.inputs.length>0||s.buttons.length>0)&&(l.p.style.marginBottom="0"),(s.inputs.length>0||s.buttons.length>0)&&(s.rtl?l.toastTexts.style.marginLeft="10px":l.toastTexts.style.marginRight="10px",s.inputs.length>0&&s.buttons.length>0&&(s.rtl?l.inputs.style.marginLeft="8px":l.inputs.style.marginRight="8px")),function(){l.toastCapsule.style.visibility="hidden",setTimeout(function(){var t=l.toast.offsetHeight,e=l.toast.currentStyle||window.getComputedStyle(l.toast),n=e.marginTop;n=n.split("px"),n=parseInt(n[0]);var o=e.marginBottom;o=o.split("px"),o=parseInt(o[0]),l.toastCapsule.style.visibility="",l.toastCapsule.style.height=t+o+n+"px",setTimeout(function(){l.toastCapsule.style.height="auto",s.target&&(l.toastCapsule.style.overflow="visible")},500),s.timeout&&i.progress(s,l.toast).start()},100)}(),function(){var t=s.position;if(s.target)l.wrapper=document.querySelector(s.target),l.wrapper.classList.add(n+"-target"),s.targetFirst?l.wrapper.insertBefore(l.toastCapsule,l.wrapper.firstChild):l.wrapper.appendChild(l.toastCapsule);else{if(-1==r.indexOf(s.position))return void console.warn("["+n+"] Incorrect position.\nIt can be › "+r);t=o||window.innerWidth<=d?"bottomLeft"==s.position||"bottomRight"==s.position||"bottomCenter"==s.position?n+"-wrapper-bottomCenter":"topLeft"==s.position||"topRight"==s.position||"topCenter"==s.position?n+"-wrapper-topCenter":n+"-wrapper-center":n+"-wrapper-"+t,l.wrapper=document.querySelector("."+n+"-wrapper."+t),l.wrapper||(l.wrapper=document.createElement("div"),l.wrapper.classList.add(n+"-wrapper"),l.wrapper.classList.add(t),document.body.appendChild(l.wrapper)),"topLeft"==s.position||"topCenter"==s.position||"topRight"==s.position?l.wrapper.insertBefore(l.toastCapsule,l.wrapper.firstChild):l.wrapper.appendChild(l.toastCapsule)}isNaN(s.zindex)?console.warn("["+n+"] Invalid zIndex."):l.wrapper.style.zIndex=s.zindex}(),function(){s.overlay&&(null!==document.querySelector("."+n+"-overlay.fadeIn")?(l.overlay=document.querySelector("."+n+"-overlay"),l.overlay.setAttribute("data-iziToast-ref",l.overlay.getAttribute("data-iziToast-ref")+","+s.ref),isNaN(s.zindex)||null===s.zindex||(l.overlay.style.zIndex=s.zindex-1)):(l.overlay.classList.add(n+"-overlay"),l.overlay.classList.add("fadeIn"),l.overlay.style.background=s.overlayColor,l.overlay.setAttribute("data-iziToast-ref",s.ref),isNaN(s.zindex)||null===s.zindex||(l.overlay.style.zIndex=s.zindex-1),document.querySelector("body").appendChild(l.overlay)),s.overlayClose?(l.overlay.removeEventListener("click",{}),l.overlay.addEventListener("click",function(t){i.hide(s,l.toast,"overlay")})):l.overlay.removeEventListener("click",{}))}(),function(){if(s.animateInside){l.toast.classList.add(n+"-animateInside");var t=[200,100,300];"bounceInLeft"!=s.transitionIn&&"bounceInRight"!=s.transitionIn||(t=[400,200,400]),s.title.length>0&&setTimeout(function(){l.strong.classList.add("slideIn")},t[0]),s.message.length>0&&setTimeout(function(){l.p.classList.add("slideIn")},t[1]),s.icon&&setTimeout(function(){l.icon.classList.add("revealIn")},t[2]);var e=150;s.buttons.length>0&&l.buttons&&setTimeout(function(){m(l.buttons.childNodes,function(t,n){setTimeout(function(){t.classList.add("revealIn")},e),e+=150})},s.inputs.length>0?150:0),s.inputs.length>0&&l.inputs&&(e=150,m(l.inputs.childNodes,function(t,n){setTimeout(function(){t.classList.add("revealIn")},e),e+=150}))}}(),s.onOpening.apply(null,[s,l.toast]);try{var b=new CustomEvent(n+"-opening",{detail:s,bubbles:!0,cancelable:!0});document.dispatchEvent(b)}catch(L){console.warn(L)}setTimeout(function(){l.toast.classList.remove(n+"-opening"),l.toast.classList.add(n+"-opened");try{var t=new CustomEvent(n+"-opened",{detail:s,bubbles:!0,cancelable:!0});document.dispatchEvent(t)}catch(e){console.warn(e)}s.onOpened.apply(null,[s,l.toast])},1e3),s.drag&&(a?(l.toast.addEventListener("touchstart",function(t){y.startMoving(this,i,s,t)},!1),l.toast.addEventListener("touchend",function(t){y.stopMoving(this,t)},!1)):(l.toast.addEventListener("mousedown",function(t){t.preventDefault(),y.startMoving(this,i,s,t)},!1),l.toast.addEventListener("mouseup",function(t){t.preventDefault(),y.stopMoving(this,t)},!1))),s.closeOnEscape&&document.addEventListener("keyup",function(t){t=t||window.event,27==t.keyCode&&i.hide(s,l.toast,"esc")}),s.closeOnClick&&l.toast.addEventListener("click",function(t){i.hide(s,l.toast,"toast")}),i.toast=l.toast},e});
